Move utils from test to test utils module
